Day 89 Morning: Order model created, admin registered and checkout view logic added

+++ store/views.py
@@ -1,10 +1,10 @@
 from django.shortcuts import render , get_object_or_404 , redirect
-from .models import Product , Category
+from .models import OrderItem, Product , Category
 from django.db.models import Q
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from .models import Cart, CartItem, Product
+from .models import Cart, CartItem, Product, Order, OrderItem
 def product_list(request):
     category_id = request.GET.get('category')
     search_query = request.GET.get('search')
@@ -57,11 +57,12 @@ def cart_detail(request):
 def decrease_cart_item(request, product_id):
     cart = Cart.objects.get(user=request.user)
     product = get_object_or_404(Product, id=product_id)
-    cart_item = CartItem.objects.get(cart=cart,product=product)
+    cart_item = CartItem.objects.filter(cart=cart,product=product).first()
 
-    if cart_item.quantity > 1:
-        cart_item.quantity -= 1
-        cart_item.save()
+    if cart_item:
+        if cart_item.quantity > 1:
+            cart_item.quantity -= 1
+            cart_item.save()
     else:
         cart_item.delete()
     return redirect('cart_detail')
@@ -69,6 +70,37 @@ def decrease_cart_item(request, product_id):
 def remove_cart_item(request, product_id):
     cart = Cart.objects.get(user=request.user)
     product = get_object_or_404(Product, id=product_id)
-    cart_item = CartItem.objects.get(cart=cart, product=product)
-    cart_item.delete()
+    cart_item = CartItem.objects.filter(cart=cart, product=product).first()
+    if cart_item:
+        cart_item.delete()
     return redirect('cart_detail')
+@login_required
+def checkout(request):
+    cart = Cart.objects.get(user=request.user)
+    if request.method == 'POST':
+        full_name = request.POST.get('full_name')
+        phone_number = request.POST.get('phone_number')
+        address = request.POST.get('address')
+
+        order = Order.objects.create(
+            user=request.user,
+            full_name=full_name,
+            phone_number=phone_number,
+            address=address,
+            total_price=cart.get_total_price(),
+            is_paid=True  # Assuming payment is successful for simplicity
+        )
+
+        for item in cart.items.all():
+            OrderItem.objects.create(
+                order=order,
+                product=item.product,
+                price=item.product.price,
+                quantity=item.quantity
+            )
+
+        cart.items.all().delete()  # Clear the cart after checkout
+        return redirect('order_success')  # Redirect to a success page after checkout
+    return render(request, 'store/checkout.html', {'cart': cart})
+    
+
